Variables for OpenAI API and Model (#1172)
Created a variable for the OpenAI Base URL and model that will be used.
This allows for utilizing other services that adhere to the OpenAI API
spec, as well as testing out additional models that work best with the
function calling and requests.

I noticed that different model versions were used throughout the code
(e.g. `gpt-3.5-turbo-1106` vs `gpt-3.5-turbo-0613`. This patch only
included a single model for all locations. If different models were used
due to performance reasons, this can be separated out.